refactor(SectioningContent): levelContext.tsをLevelContext.tsにリネーム
Contextを定義するファイルは他がいずれもパスカルケース
（AppLauncherContext.tsx・SideNavContext.tsx など）で、このファイルだけが
小文字始まりだった。

import元2箇所のパスもあわせて修正している。
